requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --format requirements.txt --output-file requirements.txt --no-hashes
blinker==1.9.0
    # via flask
cffi==2.0.0 ; platform_python_implementation != 'PyPy'
    # via cryptography
click==8.3.1
    # via flask
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   pytest
cryptography==46.0.4
    # via jwks-server
flask==3.1.2
    # via jwks-server
iniconfig==2.3.0
    # via pytest
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   werkzeug
packaging==26.0
    # via pytest
pluggy==1.6.0
    # via pytest
pycparser==3.0 ; implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'
    # via cffi
pygments==2.19.2
    # via pytest
pyjwt==2.11.0
    # via jwks-server
pytest==9.0.2
    # via jwks-server
ruff==0.15.0
werkzeug==3.1.5
    # via flask
